Moved table controller code to the main sample file
